Check if connected to database before doing entity transations

+++ services/game/src/database/entity/index.ts
@@ -1,6 +1,7 @@
 import { ObjectOfAny } from '@rogueworld/types';
 import { warning } from '@rogueworld/utils';
 import { CallbackError, Document } from 'mongoose';
+import { isDBConnected } from '..';
 import EntityModel, { EntityDocument } from './EntityModel';
 
 export { default as EntityModel, EntityDocument } from './EntityModel';
@@ -9,6 +10,8 @@ export function createEntityDocument(
     data: EntityDocument,
     onSuccess: (documentId: string) => void,
 ) {
+    if(!isDBConnected()) return;
+
     EntityModel.create(
         data,
         (err: CallbackError, res: Document) => {
@@ -23,6 +26,8 @@ export function createEntityDocument(
 }
 
 export function updateEntityDocument(id: string, data: ObjectOfAny) {
+    if(!isDBConnected()) return;
+
     EntityModel.findByIdAndUpdate(id, data, undefined, (err) => {
         if(err) {
             warning(err);
@@ -31,6 +36,8 @@ export function updateEntityDocument(id: string, data: ObjectOfAny) {
 }
 
 export function deleteEntityDocument(id: string) {
+    if(!isDBConnected()) return;
+
     EntityModel.findByIdAndDelete(id, undefined, (err) => {
         if(err) {
             warning(err);
@@ -39,6 +46,8 @@ export function deleteEntityDocument(id: string) {
 }
 
 export async function getPaginatedEntityDocuments(page: number, limit: number) {
+    if(!isDBConnected()) return;
+
     return EntityModel
         .find()
         .limit(limit)
